added '"nozzle_offsetting_for_disallowed_areas": false,' line to the metadata section
This patches the ability for the machine to think (incorrectly so) both nozzles can reach all locations on the buildplate, which acts as a manual override that allows for full buildplate surface area utilization with dual extrusion. the left extruder to print parts on the left part of the buildplate. the right extruder prints the parts on the right side of the buildplate. any attempt to print left noxxled parts on the right edge of the plate or vice versa WILL result in a XY gantry crash (the very software stop designed to prevent this from occuring is suppressed in this push so that full-buildplate dual printing is made possible)